requirements.txt
beautifulsoup4==4.13.4
fastapi==0.116.1
graphviz==0.21
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.74
langchain-openai==0.3.30
langchain-text-splitters==0.3.9
langgraph==0.6.5
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.2
langsmith==0.4.14
numpy==2.3.2
openai==1.100.1
pydantic==2.11.7
pypdf==6.0.0
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.31.0
starlette==0.47.2
streamlit==1.48.1
uvicorn==0.35.0
python-multipart==0.0.19

